add permission for user to read all user & delete

+++ internal/http/handler/user.go
@@ -41,6 +41,12 @@ func (h *UserHandler) GeneratePassword(c echo.Context) error {
 
 func (h *UserHandler) FindAllUser(c echo.Context) error {
 	users, err := h.userService.FindAll(c.Request().Context())
+	userToken := c.Get("user").(*jwt.Token)
+	claims := userToken.Claims.(*common.JwtCustomClaims)
+	// Check if the user's id same with jwt id
+	if claims.Role != "admin" {
+		return util.JSONResponse(c, http.StatusForbidden, "You don't have access to this user", nil)
+	}
 	if err != nil {
 		return util.JSONResponse(c, http.StatusInternalServerError, err.Error(), nil)
 	}
@@ -56,8 +62,10 @@ func (h *UserHandler) FindOneUser(c echo.Context) error {
 	userToken := c.Get("user").(*jwt.Token)
 	claims := userToken.Claims.(*common.JwtCustomClaims)
 	// Check if the user's id same with jwt id
-	if id != claims.ID {
-		return util.JSONResponse(c, http.StatusForbidden, "You don't have access to this user", nil)
+	if claims.Role != "admin" {
+		if id != claims.ID {
+			return util.JSONResponse(c, http.StatusForbidden, "You don't have access to this user", nil)
+		}
 	}
 	user, err := h.userService.FindOne(c.Request().Context(), id)
 	if err != nil {
@@ -107,6 +115,12 @@ func (h *UserHandler) DeleteUser(c echo.Context) error {
 	id, err := strconv.ParseInt(idStr, 10, 64)
 	if err != nil {
 		return util.JSONResponse(c, http.StatusBadRequest, "Invalid user ID", nil)
+	}
+	userToken := c.Get("user").(*jwt.Token)
+	claims := userToken.Claims.(*common.JwtCustomClaims)
+	// Check if the user's id same with jwt id
+	if claims.Role != "admin" {
+		return util.JSONResponse(c, http.StatusForbidden, "You don't have access to this user", nil)
 	}
 	if err := h.userService.Delete(c.Request().Context(), id); err != nil {
 		return util.JSONResponse(c, http.StatusInternalServerError, err.Error(), nil)
